Directly insert statistics into database
This removes the min() function from being called during collection; all
aggregation is performed during final querying for display.
